pyproject.toml
[tool.poetry]
name = "whispers"
version = "0.1.0"
description = ""
authors = [""]
license = "The Unlicense"

[tool.poetry.dependencies]
python = "^3.7"
python-telegram-bot = "^13.11"


[tool.poetry.dev-dependencies]
pylint = "^2.12.2"
black = "^22.1.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
bandit = "^1.7.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
